actually commit on a feature branch
