+++ Models/ApplicationConfig.cs
@@ -1,0 +1,6 @@
+namespace AdventureLog_API.Models;
+
+public class ApplicationConfig
+{
+    public string ConnectionString { get; set; }
+}
