+++ appsettings.Development.json
@@ -1,4 +1,5 @@
 {
+  "ConnectionString": "",
   "Logging": {
     "LogLevel": {
       "Default": "Information",
